requirements.txt
streamlit
pandas
requests
openpyxl
pathlib
streamlit-authenticator
pyyaml
bcrypt
